requirements.txt
Django>=4.2.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
feedparser>=6.0.0
django-allauth>=0.57.0
requests>=2.31.0
Pillow>=10.0.0
